dataset/pyproject.toml
[project]
name = "dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langdetect>=1.0.9",
    "parakeet-mlx>=0.4.1",
    "pyannote-audio>=4.0.2",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
]
